factorize some ic3 read/write functions
