+++ cfdtools/api.py
@@ -26,6 +26,16 @@ def fileformat_writer(name, extension):
         return thisclass
     return decorator
 
+def _printreadable(string, value):
+    if isinstance(value, (int, float, str, np.int32, np.int64)):
+        print(string+':',value)
+    elif isinstance(value, np.ndarray):
+        if value.size <= 10:
+            print(string+': ndarray',value.shape, value)
+        else:
+            print(string+': ndarray',value.shape)
+    else:
+        print(string+': '+str(type(value)))
 class api_output():
     """class to handle library outputs
     """
