+++ cfdtools/ic3/writerV2.py
@@ -1,119 +1,9 @@
 
 # Import modules
-import struct
 import numpy as np
-import numpy.ma as npma
-import collections
-import sys
 import cfdtools.api as api
 import cfdtools.meshbase._mesh as _mesh
 from cfdtools.ic3._ic3 import *
-
-struct_endian = { 'native':'@', 'little':'<', 'big':'>'}
-
-###################################################################################################
-def BinaryWrite(bfile, endian, form, varargs):
-    '''
-    Method to encapsulate the few lines necessary for the translation
-    of formatted components to a piece of formatted binary.
-    input   : handle on an open file [type file identifier]
-            format of the binary token [type string]
-            list of variables to make the bytearray
-    '''
-
-    # Assume big-endian to write the file - anyways charlesx can swap
-    form = struct_endian[endian] + form
-
-    # Create a packer
-    s = struct.Struct(form)
-    #print(len(form), form,len(varargs),':')
-    # Actually pack the string now
-    packed_ba = s.pack(*varargs)
-
-    # Try to write the bytes to the file otherwise crash
-    try:
-        bfile.write(packed_ba)
-    except IOError:
-        api.io.print('error',"Fatal error. Could not write to %s. Exiting."%(bfile.name))
-        exit()
-
-class restartSectionHeader():
-    '''
-    This class is designed to handle the header that is present
-    before every variable/section/category saved into the restart file.
-    '''
-
-    def __init__(self):
-        """
-        Initialize a section header class, as it is always structured
-        in the same way, i.e. with a name, an id, a skip bytes count,
-        an information array, and an auxiliary array needed in specific
-        cases like periodicity.
-        """
-
-        self.name   = " "*ic3_restart_codes["UGP_IO_HEADER_NAME_LEN"]
-        self.id     = np.zeros((1,), dtype=np.int32)
-        self.skip   = np.zeros((1,), dtype=np.int64)
-        self.idata  = np.zeros((8,), dtype=np.int64)
-        self.rdata  = np.zeros((16,), dtype=np.float64)
-
-        # Initialize the format string
-        self.binstr = ""
-        # Fill it with prior knowledge of its content
-        for i in range(ic3_restart_codes["UGP_IO_HEADER_NAME_LEN"]) :
-            self.binstr += "c"               # name
-        self.binstr += "i"                   # id
-        self.binstr += "q"                   # skip
-        self.binstr += "qqqqqqqq"            # idata (long long)
-        self.binstr += "dddddddddddddddd"    # rdata (double)
-
-        # Keep in mind the total byte size of the header
-        self.hsize = ic3_restart_codes["UGP_IO_HEADER_NAME_LEN"] * type2nbytes["char"] +\
-                type2nbytes["int32"] +\
-                type2nbytes["int64"] +\
-                self.idata.size * type2nbytes["int64"] +\
-                self.rdata.size * type2nbytes["float64"]
-
-    def write(self, bfile, endian):
-        """
-        Once initialization is done, this method actually writes
-        the header data from the packed binary formatted string.
-        """
-
-        # Make a list from all the arguments
-        varargs = []
-        for i in range(ic3_restart_codes["UGP_IO_HEADER_NAME_LEN"]):
-            if i < len(self.name):
-                varargs.append(bytes(self.name[i], 'utf-8'))
-            else:
-                varargs.append(b'\0')
-        # for i in varargs:
-        #     print(">",i, type(i))
-        varargs.append(self.id)
-        varargs.append(self.skip)
-        for kk in range(8):
-            varargs.append(self.idata[kk])
-        for kk in range(16):
-            varargs.append(self.rdata[kk])
-        #print(varargs)
-        # Now write everything at once
-        BinaryWrite(bfile, endian, self.binstr, varargs)
-
-    def __str__(self):
-        mystring="Header:\n"
-        mystring +="Name:%s\n"%self.name
-
-        mystring +="Id:%i\n"%self.id
-        mystring +="Skip:%i\n"%self.skip
-        mystring +="idata:("
-        for i in range(8):
-            mystring +="%i,"%(self.idata[i])
-        mystring +=")\n"
-        mystring +="rdata:("
-        for i in range(16):
-            mystring +="%f,"%(self.rdata[i])
-        mystring +=")"
-        return mystring
 
 ###################################################################################################
